Add files via upload
As technology is advancing so houses are also getting smarter. Modern houses are gradually shifting from conventional switches to a centralized control system, involving remote-controlled switches. Presently, conventional wall switches located in different parts of the house make it difficult for the user to go near them to operate. Even more, it becomes more difficult for the elderly or physically handicapped people to do so. Remote controlled home automation system provides a most modern solution.

The main objective of this project is to develop a home automation system using an Arduino board.

To achieve this objective, a scale house that captures different signals, both digital and analog, has been developed. To approach the house to a real home automation application the variables under study and control are interior temperature and lighting and movement around the house. 
The house has two main operating modes. In automatic mode, it performs the measurement and executes the control of the variables, regulating itself according to the conditions to which it is exposed. In alarm mode, it controls the parameters that assure the security of the house when proprietor is away from home. 
To capture the signals, the prototype has temperature, lighting, movement and water level sensors installed, and for the regulation and control it has a fan, some LED, an acoustic warning device and a water pump. The core is an Arduino Mega board that allows the application operation and, if it is operating remotely, orders to individually controls the different actuators.
This system is designed to assist and provide support in order to fulfill the needs of elderly and disabled in home. Automation of the surrounding environment of a modern human being allows increasing his work efficiency and comfort. There has been a significant development in the area of an individual’s routine tasks and those can be automated.